refactor(orchestrators): split reporting_orchestrator god-file into mixins
Decompose the 876-line ReportingOrchestrator into 3 cohesive mixins under
orchestrators/reporting/ (data_loading, enrichment, crew_html), each <300 lines;
the orchestrator composes them so behavior is unchanged. Updated test patch
targets that reached internal collaborators (get_generator_for_crew, Path) to
their new module paths.

Verified: 237 reporting + orchestrator tests pass, ruff + mypy clean.
